+++ test/jest/controls/control.radio.test.js
@@ -1,6 +1,7 @@
 /* eslint-disable prefer-destructuring */
 import { WUPRadioControl } from "web-ui-pack";
 import WUPBaseControl from "web-ui-pack/controls/baseControl";
+import observer from "web-ui-pack/helpers/observer";
 import { initTestBaseControl, testBaseControl } from "./baseControlTest";
 import * as h from "../../testHelper";
 
@@ -109,9 +110,12 @@ describe("control.radio", () => {
       `"<fieldset><legend><strong></strong></legend><label for="txt9"><input id="txt9" type="radio" name="txt8473" tabindex="0" autocomplete="off"><span>Harry</span></label></fieldset>"`
     );
 
-    el.$options.items = [{ text: "Helica", value: 5 }];
+    const item = { text: "Helica", value: 5 };
+    el.$options.items = [item];
     jest.advanceTimersByTime(2);
     expect(onErr).toBeCalledTimes(1); // because it doesn't fit value 11
+    expect(el.$options.items[0]).toBe(item); // nested items must be not observed
+    expect(observer.isObserved(el.$options.items)).toBe(true); // but Array items itself must be observed
   });
 
   test("$options.items is complex object", async () => {